--- a/helpers/submission_helper.rb
+++ b/helpers/submission_helper.rb
@@ -19,6 +19,7 @@ module Sinatra
       def retrieve_submissions(options)
         status = (options[:status] || "RDF").to_s.upcase
         status = "RDF" if status.eql?("READY")
+        ontology_acronym = options[:ontology]
         any = status.eql?("ANY")
         include_views = options[:also_include_views] || false
         includes, page, size, order_by, _  =  settings_params(LinkedData::Models::OntologySubmission)
@@ -32,28 +33,16 @@ module Sinatra
           submissions_query = submissions_query.where({submissionStatus: [ code: status]})
         end
 
+
+        submissions_query.where(ontology: [acronym: ontology_acronym]) if ontology_acronym
+
+
         submissions_query = apply_filters(submissions_query)
         submissions_query = submissions_query.filter(Goo::Filter.new(ontology: [:viewOf]).unbound) unless include_views
         submissions_query = submissions_query.filter(filter) if filter?
 
-        # When asking to display all metadata, we are using bring_remaining on each submission. Slower but best way to retrieve all attrs
-        if includes_param.first == :all
-          includes = [:submissionId, {:contact=>[:name, :email],
-                                      :ontology=>[:administeredBy, :acronym, :name, :summaryOnly, :ontologyType, :viewingRestriction, :acl,
-                                                                             :group, :hasDomain, :views, :viewOf, :flat, :notes, :reviews, :projects],
-                                      :submissionStatus=>[:code], :hasOntologyLanguage=>[:acronym], :metrics =>[:classes, :individuals, :properties]},
-                      :submissionStatus]
-        else
-          if includes.find{|v| v.is_a?(Hash) && v.keys.include?(:ontology)}
-           includes << {:ontology=>[:administeredBy, :acronym, :name, :viewingRestriction, :group, :hasDomain,:notes, :reviews, :projects,:acl, :viewOf]}
-          end
 
-          if includes.find{|v| v.is_a?(Hash) && v.keys.include?(:contact)}
-            includes << {:contact=>[:name, :email]}
-          end
-        end
-
-        submissions = submissions_query.include(includes)
+        submissions = submissions_query.include(submission_include_params)
         if page?
           submissions.page(page, size).all
         else