### Context

A temporary fix for https://github.com/ontoportal-lirmm/ontologies_api/issues/22 , while https://github.com/ontoportal-lirmm/goo/issues/46 is not fixed.

This PR also includes a small fix but with a big response time optimization. As the bellow benchmark shows, the endpoints `/submissions?include=all` and `/ontologies/:acronym/submissions?include=all` are now **40 times** and **6 times** faster.


Use case, a fresh database with 5 ontologies and 25 submissions:
Test | Time before optimization | Time after optimization | Improvement %
-- | -- | -- | --
get all submissions (25 submissions with 12 attributes) | 0.219019 | 0.217481 | same
get all submissions with include all (25 submissions with 119 attributes) | 10.967734 | 0.243821 | 44x faster
get ontology TEST-ONT-0 submissions (5 submissions with 12 attributes) | 0.043980 | 0.050559 | same
get ontology TEST-ONT-0 submissions with include all (5 submissions with 119 attributes) | 0.455594 | 0.067641 | 6x faster


This improvement was made by removing this no-more-needed line https://github.com/ontoportal-lirmm/ontologies_api/pull/45/files#diff-20027d6ce1351a4089be545b94f15a4f6a3f771c433ac37a70874918e06b9986L366
 
### Changes 
- Refactor and extract submission_include_params helper, to get what attribute to include in a request (https://github.com/ontoportal-lirmm/ontologies_api/pull/45/commits/62c7b0e54b1e6a517d2be1e4d2deb134625a1ade)